requirements.txt
flask~=2.2.2
newsapi-python
flask_cors
pythainlp~=3.0.8
numpy~=1.23.2
gensim
scikit-learn~=1.1.2
joblib~=1.1.0
bs4~=0.0.1
pymongo[tls,srv]
selenium~=4.4.3
webdriver_manager
fasttext~=0.9.2
requests~=2.28.1
beautifulsoup4~=4.11.1
scipy~=1.9.0
setuptools~=60.2.0